requirements.txt
# Core Dependencies
groq>=0.4.0
supabase>=2.3.0
python-dotenv>=1.0.0

# Image Processing & Vision
pillow>=10.0.0

# Embedding & Vector Search (Full ML Version)
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
requests>=2.31.0
pydantic>=2.5.0
